move item content
